requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
